requirements.txt
Django==5.2.8
drf-spectacular
djangorestframework
requests
psycopg2-binary
dotenv

